pyproject.toml
[project]
name = "python-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "django-celery-beat>=2.7.0",
    "django-q2>=1.8.0",
    "django-redis>=5.4.0",
    "logfire[django]>=3.5.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pytz>=2025.1",
]

[dependency-groups]
dev = [
    "gunicorn>=23.0.0",
    "setuptools>=80.9.0",
    "uvicorn>=0.35.0",
]
